Fix dangerous casts
On Windows 64-bit, the size of long is 32-bit, thus any pointer to long cast is
not valid. In any case, one should use intptr_t and ptrdiff_t when casting
to integers. This commit attempts to fix all instances reported by GCC.
When relevant, I replaced code by the macros PTR_ADD, ALIGN_UP from system.h

+++ apps/plugins/lib/simple_viewer.c
@@ -69,7 +69,7 @@ static const char* get_next_line(const char *text, struct view_info *info)
         w = 1;
 #else
         unsigned short ch;
-        n = ((long)rb->utf8decode(ptr, &ch) - (long)ptr);
+        n = ((intptr_t)rb->utf8decode(ptr, &ch) - (intptr_t)ptr);
         if (rb->is_diacritic(ch, NULL))
             w = 0;
         else
